+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version header format: `## [x.x.x] - yyyy-mm-dd`
 
 ### Added
 
+- cmd/workflow-engine for cli
 - pkg/environments
 - pkg/jobs
 - pkg/pipelines